requirements.txt
flask
flask-cors
pillow
pytesseract
python-docx
pypdf
gunicorn